cuda-backend/pyproject.toml
[tool.poetry]
name = "chategw-api"
version = "0.1.0"
description = ""
authors = ["Kyrylo Kobets <mjr@itage.biz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = { extras = ["standard"], version = "^0.24" }
torch = { version = ">=2.0.0, !=2.0.1, !=2.1.0" }
fastapi = "^0.104"
tqdm = "^4.66.1"
sentence-transformers = "^2.2.2"
spacy = { version = "^3.7.2", extras = ["cuda12x"] }
semantic-split = "^0.1.0"
lingua-language-detector = "^1.3.3"
sacremoses = "^0.1.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
